refactor: formatSessionElapsed / sessionProjectName を shared に集約
desktop・mobile で重複していた formatElapsed / resolveProjectName を
shared/types.ts の formatSessionElapsed / sessionProjectName として統合し、
両ファイルから参照する。
